+++ src/Products/MeetingCommunes/tests/testCustomWorkflows.py
@@ -143,8 +143,8 @@ class testCustomWorkflows(MeetingCommunesTestCase):
         self.assertEqual(item.query_state(), 'accepted')
         self.assertEqual(meeting.query_state(), 'closed')
 
-    def test_pm_AdviceCustomWorkflowAdviceAwareCorrectlyUpdated(self):
-        """When an advice aware advice using a custom workflow is in a custom review_state,
+    def test_pm_AdviceCustomWorkflowDelayAwareCorrectlyUpdated(self):
+        """When an delay aware advice using a custom workflow is in a custom review_state,
            it is correctly updated by the night task."""
         self._configureFinancesAdvice()
         cfg = self.meetingConfig
@@ -199,3 +199,34 @@ class testCustomWorkflows(MeetingCommunesTestCase):
         # after update by @@pm-night-tasks, 9 days left
         self.portal.restrictedTraverse('@@pm-night-tasks')()
         self.assertEqual(item.adviceIndex[self.vendors_uid]['delay_infos']['left_delay'], 9)
+
+    def test_pm_Show_advice_on_final_wf_transition_when_item_in_advice_not_giveable_state(self):
+        """Test especially that if a finances advice is taken back in a state
+           where it is no more giveable, it is not shown if advice WF was not ended."""
+        self._configureFinancesAdvice()
+        cfg = self.meetingConfig
+        cfg.setItemAdviceStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceEditStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceViewStates((self._stateMappingFor('itemcreated'), ))
+        # create item and ask advice
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem',
+                           title="Item to advice",
+                           category='development',
+                           optionalAdvisers=(self.vendors_uid, ))
+        self.changeUser('pmReviewer2')
+        vendors_advice_portal_type = item.adapted()._advicePortalTypeForAdviser(self.vendors_uid)
+        cfg.setDefaultAdviceHiddenDuringRedaction([vendors_advice_portal_type])
+        vendors_advice = self.addAdvice(
+            item,
+            **{'advice_group': self.vendors_uid,
+               'advice_type': u'positive_with_remarks',
+               'advice_comment': u'My comment',
+               'advice_portal_type': vendors_advice_portal_type})
+        self.assertEqual(vendors_advice.query_state(), 'proposed_to_financial_manager')
+        self.assertTrue(vendors_advice.advice_hide_during_redaction)
+        # propose item, it will still be hidden during redaction
+        self.proposeItem(item)
+        self.assertEqual(vendors_advice.query_state(), 'advice_given')
+        self.assertTrue(vendors_advice.advice_hide_during_redaction)
+        self.assertTrue(item.adviceIndex[self.vendors_uid]['hidden_during_redaction'])